pyproject.toml
[tool.poetry]
name = "environmental-weather-dashboard"
version = "0.1.0"
description = ""
authors = ["VICTORS_LAPTOP\\saarv <saarms5@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
streamlit = "^1.53.1"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
pandas-stubs = "~=2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
